% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median.R
\name{mean_median}
\alias{mean_median}
\title{Mean and Median}
\usage{
mean_median(data)
}
\arguments{
\item{data}{is the numeric data or dataset thaat we will use}

\item{mean}{is the calculated mean of the data}

\item{median}{is the calculated medican of the data}
}
\value{
the new phase with mean and median
}
\description{
this function is designed to calcuate and display the mean and median at thes same time
}
\examples{
mean_median_body_mass<-mean_median(penguins$body_mass_g)
}
